st/oxr: Preserve which path the app suggested
The OXR_DEBUG_BINDINGS output uses this to print debug info.

Otherwise there should be no difference - the entire problem was that
parents like /trigger are added to each of the oxr_bindings that respresent
a component like /trigger/click, /trigger/value, etc.

So as long as we only add the act_key to the oxr_binding that we want
selected, the preferred_binding_path_index is purely informative of what
the app originally suggested.